+++ jsonargparse/_signatures.py
@@ -2,7 +2,6 @@
 
 import dataclasses
 import inspect
-import os
 import re
 from argparse import SUPPRESS, ArgumentParser
 from collections.abc import Callable
@@ -12,14 +11,12 @@ from ._actions import _ActionConfigLoad
 from ._common import (
     LoggerProperty,
     get_generic_origin,
-    get_parsing_setting,
     get_unaliased_type,
     is_final_class,
     is_subclass,
     is_subclasses_disabled,
 )
-from ._deprecated import deprecation_warning, renamed_parameter_warning
-from ._instantiation import get_class_instantiator
+from ._instantiation import dynamic_class_instantiator
 from ._namespace import Namespace, get_value_and_parent
 from ._optionals import attrs_support, get_doc_short_description, is_attrs_class, is_pydantic_model
 from ._parameter_resolvers import ParamData, get_parameter_origins, get_signature_parameters
@@ -58,7 +55,6 @@ def validate_fail_untyped(fail_untyped) -> None:
 class SignatureArguments(LoggerProperty):
     """Methods to add arguments based on signatures to an :class:`ArgumentParser` instance."""
 
-    @renamed_parameter_warning({"theclass": "class_type"})
     def add_class_arguments(
         self,
         class_type: type,
@@ -150,7 +146,6 @@ class SignatureArguments(LoggerProperty):
 
         return added_args
 
-    @renamed_parameter_warning({"theclass": "class_type", "themethod": "method_name"})
     def add_method_arguments(
         self,
         class_type: type,
@@ -383,23 +378,12 @@ class SignatureArguments(LoggerProperty):
             annotation = unvalidatable_replaced
         if default == inspect_empty:
             default = param.default
-            if default == inspect_empty:
-                if is_optional(annotation):
-                    if os.environ.get("JSONARGPARSE_DEPRECATION_WARNINGS", "").lower() == "all":
-                        deprecation_warning(
-                            "signature_optional_parameter_without_default",
-                            "Optional type parameters without a default are currently not required. "
-                            "In v5 they will be required.",
-                            stacklevel=4,
-                        )
-                    unset_sentinel = get_parsing_setting("unset_sentinel")
-                    default = unset_sentinel if unset_sentinel is not None else None
-                elif get_typehint_origin(annotation) in not_required_types:
-                    default = SUPPRESS
-                    self.logger.debug(
-                        f'Parameter "{name}" from "{src}" is NotRequired and does not have a default, '
-                        "so it is not included in the parsed namespace unless given."
-                    )
+            if default == inspect_empty and get_typehint_origin(annotation) in not_required_types:
+                default = SUPPRESS
+                self.logger.debug(
+                    f'Parameter "{name}" from "{src}" is NotRequired and does not have a default, '
+                    "so it is not included in the parsed namespace unless given."
+                )
         # Determine argument characteristics based on parameter kind and default value
         if kind == kinds.POSITIONAL_ONLY:
             is_required = True  # Always required
@@ -418,17 +402,7 @@ class SignatureArguments(LoggerProperty):
             # are meant to agree with the requiredness that the signature itself defines.
             annotation = strip_required_typehint(annotation, is_required, f'parameter "{name}" from "{src}"')
         if is_required and annotation == inspect_empty and fail_untyped is False:
-            if os.environ.get("JSONARGPARSE_DEPRECATION_WARNINGS", "").lower() == "all":
-                deprecation_warning(
-                    "fail_untyped_false_required_parameter",
-                    "With fail_untyped=False, required parameters without a type annotation are currently "
-                    "set to optional with default None. In v5 the type will be set to Untyped but the "
-                    "parameter will remain required.",
-                    stacklevel=4,
-                )
             annotation = Untyped
-            default = None
-            is_required = False
         is_required_link_target = False
         if is_required and linked_targets is not None and name in linked_targets:
             default = None
@@ -677,8 +651,7 @@ def group_instantiate_class(group, cfg):
         value = {}
         parent = cfg
         key = group.dest
-    instantiator_fn = get_class_instantiator()
-    parent[key] = instantiator_fn(group.group_class, **value)
+    parent[key] = dynamic_class_instantiator(group.group_class, **value)
 
 
 def strip_title(value):