+++ jsonargparse_tests/__main__.py
@@ -1,5 +1,4 @@
 """Run all unit tests in package."""
-# pragma: no cover
 
 import os
 import sys