+++ jsonargparse_tests/test_typehints.py
@@ -96,6 +96,11 @@ def test_add_argument_given_type_and_null_action(parser):
     assert parser.get_defaults().op1 is None
 
 
+def test_add_argument_nargs_zero_not_allowed(parser):
+    with pytest.raises(ValueError, match="does not allow nargs=0"):
+        parser.add_argument("--op1", type=int, nargs=0)
+
+
 @pytest.mark.parametrize("typehint", [Namespace, Optional[Namespace], Union[int, Namespace], List[Namespace]])
 def test_namespace_unsupported_as_type(parser, typehint):
     with pytest.raises(ValueError, match="Namespace .* not supported as a type"):